pyproject.toml
[tool.poetry]
name = "aitrix_testcase"
version = "0.1.0"
description = ""
authors = ["Akim Zholdoshbek <akim.j.2002@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
gunicorn = "^20.1.0"
psycopg2 = "^2.9.3"
pip = "^22.0.4"
django-filter = "^21.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
